Refactor documentation into separate pages with shared sidebar
- Split monolithic index.php into 18 individual documentation pages
- Create reusable sidebar.php component imported by each page
- Add docs-header.php and docs-footer.php for common page structure
- Update index.php as landing page with section overview cards
- Add page navigation buttons (prev/next) on each content page
- Simplify main.js for page-based navigation (removed scroll logic)
- Update search.js with static page index for multi-page search
- Add CSS for section link cards and page navigation buttons
