+++ documentation/docs-footer.php
@@ -1,0 +1,8 @@
+    </div>
+
+    <footer class="footer">
+        <div id="includeFooter"></div>
+    </footer>
+</body>
+
+</html>
