+++ documentation/installation.php
@@ -1,0 +1,45 @@
+<?php
+$pageTitle = 'Installation Guide';
+$pageDescription = 'Learn how to download and install Argo Books on your Windows, macOS, or Linux computer.';
+$currentPage = 'installation';
+
+include 'docs-header.php';
+include 'sidebar.php';
+?>
+
+        <!-- Main Content -->
+        <main class="content">
+            <section class="article">
+                <h1>Installation Guide</h1>
+                <p>Follow these steps to install Argo Books on your computer.</p>
+
+                <ol class="steps-list">
+                    <li>Download the installer <a class="link" href="../downloads">here</a></li>
+                    <li>Run the installer file (Argo Books Installer.exe)</li>
+                    <li>Follow the installation wizard</li>
+                    <li>Launch Argo Books from your desktop or start menu</li>
+                </ol>
+
+                <div class="info-box">
+                    <strong>Tip:</strong> Your computer may display a security warning. This is normal for new applications.
+                    Click "More info" and then "Run anyway" to proceed.
+                </div>
+
+                <div class="page-navigation">
+                    <a href="system-requirements.php" class="nav-button prev">
+                        <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
+                            <path d="M15 18l-6-6 6-6"></path>
+                        </svg>
+                        Previous: System Requirements
+                    </a>
+                    <a href="quick-start.php" class="nav-button next">
+                        Next: Quick Start Tutorial
+                        <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
+                            <path d="M9 18l6-6-6-6"></path>
+                        </svg>
+                    </a>
+                </div>
+            </section>
+        </main>
+
+<?php include 'docs-footer.php'; ?>
