+++ documentation/product-management.php
@@ -1,0 +1,93 @@
+<?php
+$pageTitle = 'Product Management';
+$pageDescription = 'Learn how to manage products and categories in Argo Books. Create categories, add products, and organize your inventory.';
+$currentPage = 'product-management';
+
+include 'docs-header.php';
+include 'sidebar.php';
+?>
+
+        <!-- Main Content -->
+        <main class="content">
+            <section class="article">
+                <h1>Product Management</h1>
+                <p>Learn how to create categories and manage products in Argo Books.</p>
+
+                <h2>Creating Categories to organize your products</h2>
+                <ol class="steps-list">
+                    <li>Go to "Manage Categories" in the top menu
+                    <li>Choose whether the category is for purchases or sales
+                    <li>Enter the category name
+                    <li>Click "Add Category"
+                </ol>
+
+                <h2>Adding Products</h2>
+                <ol class="steps-list">
+                    <li>Go to "Manage Products" in the top menu
+                    <li>Select whether the product is for purchases or sales
+                    <li>Enter the product ID and name
+                    <li>Select a category to keep things organized
+                    <li>Enter country and company of origin
+                    <li>Click "Add Product"
+                </ol>
+
+                <div class="info-box">
+                    <p>Let's say you run a t-shirt store. Here's how you might set up three products within 2
+                        categories:
+                    </p>
+
+                    <table class="comparison-table">
+                        <tr>
+                            <th>Category</th>
+                            <th>Product ID</th>
+                            <th>Product Name</th>
+                            <th>Country</th>
+                            <th>Company</th>
+                        </tr>
+                        <tr>
+                            <td>Men's T-Shirts</td>
+                            <td>TS001</td>
+                            <td>Black T-Shirt</td>
+                            <td>United States</td>
+                            <td>Cotton Mills Ltd</td>
+                        </tr>
+                        <tr>
+                            <td>Men's T-Shirts</td>
+                            <td>TS002</td>
+                            <td>White Linen T-Shirt</td>
+                            <td>Italy</td>
+                            <td>FabricCo</td>
+                        </tr>
+                        <tr>
+                            <td>Women's T-Shirts</td>
+                            <td>TS003</td>
+                            <td>White T-Shirt</td>
+                            <td>Germany</td>
+                            <td>TextileCo</td>
+                        </tr>
+                    </table>
+                </div>
+
+                <div class="warning-box">
+                    <strong>Important:</strong> Free version users are limited to 10 products. Upgrade to the paid
+                    version <a class="link" href="../upgrade/">here</a> for unlimited products.
+                </div>
+
+                <div class="page-navigation">
+                    <a href="version-comparison.php" class="nav-button prev">
+                        <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
+                            <path d="M15 18l-6-6 6-6"></path>
+                        </svg>
+                        Previous: Free vs. Paid Version
+                    </a>
+                    <a href="sales-tracking.php" class="nav-button next">
+                        Next: Purchase/Sales Tracking
+                        <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
+                            <path d="M9 18l6-6-6-6"></path>
+                        </svg>
+                    </a>
+                </div>
+            </section>
+        </main>
+
+<?php include 'docs-footer.php'; ?>
