+++ documentation/quick-start.php
@@ -1,0 +1,46 @@
+<?php
+$pageTitle = 'Quick Start Tutorial';
+$pageDescription = 'Get started quickly with Argo Books. Learn the basic steps to set up your business and start tracking products and sales.';
+$currentPage = 'quick-start';
+
+include 'docs-header.php';
+include 'sidebar.php';
+?>
+
+        <!-- Main Content -->
+        <main class="content">
+            <section class="article">
+                <h1>Quick Start Tutorial</h1>
+                <p>Get up and running with Argo Books in just a few steps.</p>
+
+                <ol class="steps-list">
+                    <li>Choose your default currency and create your first company</li>
+                    <li>Add accountants and companies you will be working with</li>
+                    <li>Set up categories to organize your products</li>
+                    <li>Add your initial products</li>
+                    <li>Add purchases and sales</li>
+                </ol>
+
+                <div class="info-box">
+                    <strong>Note:</strong> A video tutorial is coming soon to help you visualize these steps. Check back
+                    later for a comprehensive walkthrough.
+                </div>
+
+                <div class="page-navigation">
+                    <a href="installation.php" class="nav-button prev">
+                        <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
+                            <path d="M15 18l-6-6 6-6"></path>
+                        </svg>
+                        Previous: Installation Guide
+                    </a>
+                    <a href="version-comparison.php" class="nav-button next">
+                        Next: Free vs. Paid Version
+                        <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2">
+                            <path d="M9 18l6-6-6-6"></path>
+                        </svg>
+                    </a>
+                </div>
+            </section>
+        </main>
+
+<?php include 'docs-footer.php'; ?>
