+++ documentation/style.css
@@ -80,27 +80,9 @@ body {
   font-weight: 500;
 }
 
-/* Style for clickable nav items */
+/* Style for nav items */
 .nav-links li {
   margin-bottom: 0.2rem;
-  cursor: pointer;
-  font-size: 1rem;
-  display: block;
-  padding: 0.3rem 0.5rem;
-  border-radius: 0.375rem;
-  transition: all 0.2s;
-}
-
-.nav-links li:hover {
-  background: var(--gray-100);
-  color: var(--primary-blue);
-}
-
-/* Style for active (current) nav item */
-.nav-links li.active {
-  background: var(--blue-100);
-  color: var(--primary-blue);
-  font-weight: 500;
 }
 
 /* Mobile menu toggle button */
@@ -517,5 +499,114 @@ body {
 @media (max-width: 768px) {
   .requirement-card {
     padding: 20px;
+  }
+}
+
+/* Section link cards for landing page */
+.section-links {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(280px, 1fr));
+  gap: 1rem;
+  margin-top: 1rem;
+}
+
+.section-link-card {
+  display: block;
+  padding: 1.25rem;
+  background: var(--whiteBackground);
+  border: 1px solid var(--gray-border);
+  border-radius: 0.5rem;
+  text-decoration: none;
+  transition: all 0.2s ease;
+}
+
+.section-link-card:hover {
+  border-color: var(--primary-blue);
+  box-shadow: 0 4px 12px rgba(37, 99, 235, 0.15);
+  transform: translateY(-2px);
+}
+
+.section-link-card h3 {
+  font-size: 1.1rem;
+  color: var(--gray-900);
+  margin: 0 0 0.5rem 0;
+}
+
+.section-link-card p {
+  font-size: 0.9rem;
+  color: var(--gray-600);
+  margin: 0;
+  line-height: 1.4;
+}
+
+/* Contact box */
+.contact-box {
+  background: var(--blue-100);
+  border: 1px solid var(--primary-blue);
+  border-radius: 0.5rem;
+  padding: 1rem;
+  margin-top: 1rem;
+}
+
+.contact-box p {
+  margin: 0;
+}
+
+/* Page navigation buttons */
+.page-navigation {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-top: 2rem;
+  padding-top: 1.5rem;
+  border-top: 1px solid var(--gray-border);
+  gap: 1rem;
+}
+
+.nav-button {
+  display: inline-flex;
+  align-items: center;
+  gap: 0.5rem;
+  padding: 0.75rem 1.25rem;
+  background: var(--whiteBackground);
+  border: 1px solid var(--gray-border);
+  border-radius: 0.5rem;
+  text-decoration: none;
+  color: var(--gray-700);
+  font-size: 0.9375rem;
+  transition: all 0.2s ease;
+}
+
+.nav-button:hover {
+  border-color: var(--primary-blue);
+  color: var(--primary-blue);
+  background: var(--blue-100);
+}
+
+.nav-button.prev {
+  margin-right: auto;
+}
+
+.nav-button.next {
+  margin-left: auto;
+}
+
+.nav-button svg {
+  flex-shrink: 0;
+}
+
+@media (max-width: 600px) {
+  .page-navigation {
+    flex-direction: column;
+  }
+
+  .nav-button {
+    width: 100%;
+    justify-content: center;
+  }
+
+  .nav-button.prev,
+  .nav-button.next {
+    margin: 0;
   }
 }
